Clean up HIP-869 tests
