+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/spec/transactions/node/HapiNodeCreate.java
@@ -130,9 +130,7 @@ public class HapiNodeCreate extends HapiTxnOp<HapiNodeCreate> {
     }
 
     private void genKeysFor(final HapiSpec spec) {
-        if (adminKeyName.isPresent() || adminKeyShape.isPresent()) {
-            adminKey = netOf(spec, adminKeyName, adminKeyShape);
-        }
+        adminKey = netOf(spec, adminKeyName, adminKeyShape);
     }
 
     @Override
