+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/hip869/NodeCreateTest.java
@@ -14,7 +14,13 @@
  * limitations under the License.
  */
 
-package com.hedera.services.bdd.suites.addressbook;
+package com.hedera.services.bdd.suites.hip869;
+
+import com.hedera.services.bdd.junit.HapiTest;
+import org.junit.jupiter.api.DynamicTest;
+import org.junit.jupiter.api.Tag;
+
+import java.util.stream.Stream;
 
 import static com.hedera.services.bdd.junit.TestTags.EMBEDDED;
 import static com.hedera.services.bdd.spec.HapiSpec.defaultHapiSpec;
@@ -26,22 +32,17 @@ import static com.hedera.services.bdd.spec.utilops.EmbeddedVerbs.viewNode;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.newKeyNamed;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.validateChargedUsdWithin;
 import static com.hedera.services.bdd.suites.HapiSuite.ONE_HBAR;
+import static com.hedera.services.bdd.suites.HapiSuite.ONE_HUNDRED_HBARS;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.BUSY;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INSUFFICIENT_TX_FEE;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.UNAUTHORIZED;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import com.hedera.services.bdd.junit.HapiTest;
-import java.util.stream.Stream;
-import org.junit.jupiter.api.DynamicTest;
-import org.junit.jupiter.api.Tag;
-
-public class NodeCreateSuite {
+public class NodeCreateTest {
     @HapiTest
     @Tag(EMBEDDED)
     final Stream<DynamicTest> createNodeWorks() {
         final String description = "His vorpal blade went snicker-snack!";
-
         return hapiTest(
                 nodeCreate("ntb").description(description),
                 viewNode(
@@ -126,20 +127,15 @@ public class NodeCreateSuite {
     }
 
     @HapiTest
-    @Tag(EMBEDDED)
     final Stream<DynamicTest> failsAtIngestForUnAuthorizedTxns() {
         final String description = "His vorpal blade went snicker-snack!";
-        return defaultHapiSpec("failsAtIngestForUnAuthorizedTxns")
-                .given(
-                        cryptoCreate("payer").balance(10_000_000_000L),
-                        nodeCreate("ntb")
-                                .payingWith("payer")
-                                .signedBy("payer")
-                                .description(description)
-                                .fee(ONE_HBAR)
-                                .hasPrecheck(BUSY)
-                                .via("nodeCreation"))
-                .when()
-                .then();
+        return hapiTest(
+                cryptoCreate("payer").balance(ONE_HUNDRED_HBARS),
+                nodeCreate("ntb")
+                        .payingWith("payer")
+                        .description(description)
+                        .fee(ONE_HBAR)
+                        .hasPrecheck(BUSY)
+                        .via("nodeCreation"));
     }
 }
